feat(format): allow reading nested json attributes
It is now possible to access nested attributes in a JSON object by
separating the attribute names with slashes in the JSON tag of a Field
object.

Consider this JSON object:
{
  "id": 1234,
  "profile": {
    "name": "foo"
  }
}
As always to access the ID, set the jsonTag to "id".
To get the profile's name, set it to "profile/name".